Finish high score scree. Add game over screen with name input.
Text rendered in high scores screen except title has no kerning, the
same as in game over screen except title.

Name input in game over screen uses up/down to cycle through big letter
english characters for 3 characters in length. Enter to submit.

Also fixed GameStateManager class to call dispose when new state is set.

+++ core/src/io/wasin/asteroids/data/PlayerScore.kt
@@ -11,5 +11,5 @@ package io.wasin.asteroids.data
  * Thus this will have major effect in production environment
  */
 
-data class PlayerScore(var name: String="---", var score: Int=0)
+data class PlayerScore(var name: String="---", var score: Long=0)
 data class PlayerSave(var highScores: Array<PlayerScore> = emptyArray())
